add guard so no exited processes are killed

+++ resources/js/electron-plugin/src/index.ts
@@ -278,15 +278,18 @@ class NativePHP {
       .filter((p) => p !== undefined)
       .forEach((process) => {
         if (!process || !process.pid) return;
+        if (process.killed && process.exitCode !== null) return;
 
         try {
-          process.kill(0); // Test if process exists (throws error if not)
-          // @ts-ignore
-          killSync(process.pid, 'SIGTERM', true); // Kill tree
-          ps.kill(process.pid); // Sometimes does not kill the subprocess of php server
+
+            // @ts-ignore
+            killSync(process.pid, 'SIGTERM', true); // Kill tree
+            ps.kill(process.pid); // Sometimes does not kill the subprocess of php server
+
         } catch (err) {
           console.error(err);
         }
+
       });
   }
 }
